requirements.txt
gymnasium[box2d]
torch
torchsummary
tqdm
numpy
Pillow
tensorboard
ipykernel
